Classe AddressService atualizada e arrumada, tal qual classe AddressDTO2 refatorada para AddressDTO.
